+++ src/Loan.h
@@ -5,30 +5,30 @@
 
 class Loan {
 private:
-    float amount;
-    float interest;
-    int tenure;  // in months
-    float emi;
+    long double amount;
+    long double interest;
+    long int tenure;  // in months
+    long double emi;
     
 public:
     Loan();
-    Loan(float amt, float rate, int months);
+    Loan(long double amt, long double rate, long int months);
     
     // setters
-    void setAmount(float amt);
-    void setInterest(float rate);
-    void setTenure(int months);
+    void setAmount(long double amt);
+    void setInterest(long double rate);
+    void setTenure(long int months);
     
     // getters
-    float getAmount();
-    float getInterest();
-    int getTenure();
-    float getEMI();
+    long double getAmount();
+    long double getInterest();
+    long int getTenure();
+    long double getEMI();
     
     // calculation
     void calculateEMI();
-    float getTotalPayment();
-    float getTotalInterest();
+    long double getTotalPayment();
+    long double getTotalInterest();
     void displayDetails();
 };
